test(ingest): hoist temp-workspace boilerplate into shared fixture
CI's fallow with auto-changed-since vs main flagged a 27-line / 57-token
clone group between the two ingest CLI test files (the tempDirs +
afterEach + makeWorkspace boilerplate they each redeclared). Local
fallow missed it because it didn't see both files as changed.

Extract useIngestWorkspaces() into test/fixtures/ingest-workspace.ts
and migrate both files to use it. Same pattern as useAimockLifecycle.
